pyproject.toml
[project]
name = "agent-joao"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.11"
dependencies = [
    "fastapi>=0.128.0",
    "langchain[openai]>=1.2.7",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
